fix: filters closing on the first select
